Fully working. Fixed most bugs
