+++ src/world/generator/Flat.js
@@ -1,5 +1,3 @@
-const ServerLevelChunkPacket = require("../../network/packets/ServerLevelChunkPacket");
-
 const WorldGenerators = require("../types/WorldGenerators");
 
 const Generator = require("./Generator");
@@ -18,9 +16,10 @@ class Flat extends Generator {
 	 * @returns {Buffer} 
 	 */
 	getChunkData() {
-		chunkData = Buffer.alloc(16 * 16 * 256);
-		for (let y = 0; y < 256; y++) {
-			for (let x = 0; x < 16; x++) {
+		chunkData = Buffer.alloc(16 * 256 * 16);
+
+		for (let x = 0; x < 16; x++) {
+			for (let y = 0; y < 256; y++) {
 				for (let z = 0; z < 16; z++) {
 					const index = y * 16 * 16 + z * 16 + x;
 
@@ -38,22 +37,6 @@ class Flat extends Generator {
 		}
 
 		return chunkData
-	}
-
-	generate(player) {
-		const chunkRadius = player.world.getChunkRadius();
-
-		for (let x = player.location.x - chunkRadius; x <= player.location.x + chunkRadius; x++) {
-			for (let z = player.location.z - chunkRadius; z <= player.location.z + chunkRadius; z++) {
-				const levelChunk = new ServerLevelChunkPacket();
-				levelChunk.setX(x);
-				levelChunk.setZ(z);
-				levelChunk.setSubChunkCount(1);
-				levelChunk.setCacheEnabled(false);
-				levelChunk.setPayload(this.getChunkData());
-				levelChunk.writePacket(player);
-			}
-		}
 	}
 }
