+++ src/world/generator/Generator.js
@@ -15,6 +15,8 @@
  */
 /* eslint-disable no-unused-vars */
 
+const ServerLevelChunkPacket = require('../../network/packets/ServerLevelChunkPacket');
+
 class Generator {
 	/**
 	 * Returns the generator name
@@ -37,8 +39,20 @@ class Generator {
 	/**
 	 * Generates chunks
 	 */
-	generate() {
-		return null;
+	generate(player) {
+		const chunkRadius = player.world.getChunkRadius();
+
+        for (let x = player.location.x - chunkRadius; x <= player.location.x + chunkRadius; x++) {
+            for (let z = player.location.z - chunkRadius; z <= player.location.z + chunkRadius; z++) {
+                const levelChunk = new ServerLevelChunkPacket();
+                levelChunk.setX(x);
+                levelChunk.setZ(z);
+                levelChunk.setSubChunkCount(1);
+                levelChunk.setCacheEnabled(false);
+                levelChunk.setPayload(this.getChunkData());
+                levelChunk.writePacket(player);
+            }
+        }
 	}
 }
