+++ src/world/generator/Void.js
@@ -1,10 +1,6 @@
-const ServerLevelChunkPacket = require("../../network/packets/ServerLevelChunkPacket");
-
 const WorldGenerators = require("../types/WorldGenerators");
 
 const Generator = require("./Generator");
-
-let chunkData;
 
 class Flat extends Generator {
     /**
@@ -18,25 +14,7 @@ class Flat extends Generator {
      * @returns {Buffer} 
      */
     getChunkData() {
-        chunkData = Buffer.alloc(16 * 16 * 256).fill(0);
-
-        return chunkData
-    }
-
-    generate(player) {
-        const chunkRadius = player.world.getChunkRadius();
-
-        for (let x = player.location.x - chunkRadius; x <= player.location.x + chunkRadius; x++) {
-            for (let z = player.location.z - chunkRadius; z <= player.location.z + chunkRadius; z++) {
-                const levelChunk = new ServerLevelChunkPacket();
-                levelChunk.setX(x);
-                levelChunk.setZ(z);
-                levelChunk.setSubChunkCount(1);
-                levelChunk.setCacheEnabled(false);
-                levelChunk.setPayload(this.getChunkData());
-                levelChunk.writePacket(player);
-            }
-        }
+        return Buffer.alloc(16 * 256 * 16).fill(0)
     }
 }
